requirements.txt
astroid==2.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.1
EasyProcess==0.2.5
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mt-940==4.13.0
oauthlib==3.0.1
psycopg2==2.7.6.1
pylint==2.2.2
python-decouple==3.1
pytz==2018.7
requests==2.21.0
requests-oauthlib==0.4.2
sepaxml==2.0.0
six==1.12.0
urllib3==1.24.1
webencodings==0.5.1
wrapt==1.11.1
